fix(lru): segfault in LruCache.pop
the problem was that it would deinit the node before returning node data. this almost always worked fine because you would need the allocator to reuse or unmap the memory to actually see a problem, and this was typically not happening before the return statement on the next line of code was executed. but on rare occasion the address would be unmapped, causing a segfault.

the solution is to defer the deinit so it happens after the return statement copies the value