fix(bincode): deserialization failure (ShredType)
All enums are deserialized with the assumption they have a size of 32 bits in bincode. ShredType however is a struct in rust and enum in zig. It is serialized as 8 bits. This causes deserialization errors reading the data from mainnet because it grabs 32 bits and reads a much larger number than any of the enum variants in zig.

Add logic to bincode implementation that looks for a declaration called BincodeSize within the enum. If it exists, use that instead of u32. Set it to u8 for ShredType.