fix(gossip): #50 Double free in ActiveSet
# Root cause
Multiple contacts with the same pubkey were being passed to the rotate method. This resulted in duplicate items being included in the array, but the hashmap internally prevents there from being duplicates. So the next time rotate is called, it would iterate through the array and attempt to deinit the same hashmap entry twice.

# Solution
Duplicates are detected and skipped during the insertion loop by checking for their presence in the hashmap.

# Alternate Solution
This could have also been addressed with a simpler approach: Remove the peers field from ActiveSet. This field does not seem necessary, since the same data is already in the hashmap.

Even though it would be simpler code, it is a more significant change in behavior and compatibility. To be safe, I decided not to take this approach.