refactor(prometheus): more straightforward data ownership
metrics structs can use normal initialization so they aren's so esoteric. the registry is the context where we'd like to allocate everything in an arena and manage the memory centrally. so, for clarity, that scope can also be in charge of allocating the structs. that way, the structs can behave as normal structs on their own, without requiring special memory management