fix(gossip): empty crds table bug -> increase removeOldLabels timeout to epoch duration
crds values are typically supposed to have a timeout of the epoch duration, unless it's from a node with zero stake weight, then it should timeout after 15 seconds

we don't track stake weight, so we can't do this conditional logic. the existing compromises by using the 15 second timeout for all crds values.

it's not practical to purge all crds values after 15 seconds. this causes the crds table to regularly be emptied.

instead, this change applies an epoch-long timeout to every value. this is not perfect because 0-stake nodes are overly trusted. for now we can live with the noise to ensure we get some meaningful signal. later, we can apply more aggressive timeouts to filter out values received from nodes with 0 stake.