fix(gossip): CrdsData use-after-free segfault
slices in CrdsData are freed after processing the incoming message, which causes a segfault when the messages need to be serialized and sent to other peers. this change instead only "shallow" frees the protocol message, leaving pointers in the CrdsData intact. The CrdsData pointers are then freed only when removed from the CrdsTable

this is a quick and dirty solution for two reasons:

1. it is theoretically possible that a CrdsData could be copied from the CrdsTable (including its contained pointers). then the data could be removed, leading to the pointers being freed. then the copied version of the pointer (now dangling) could attempt a dereference, resulting in a segfault. it does not appear that this happens in the code currently but it would be very easy to accidentally add this bug. it would be preferable if these data structures ensured the pointer is never dangling
2. it is extra maintenance burden to be sure that shallowFree is properly implemented at all times. it would be very easy to overlook this method when making changes to the contained data structures.