fix(gossip): crds remove errors from flawed contact info filtering
when removing old labels, there is a possibility for a segfault or a double free. this commit does not resolve the memory unsoundness, but it avoids the circumstance that triggers the problem.

a recent change started tracking a node's shred version from not just LegacyContactInfo but also ContactInfo. this means that some items were allowed into the crds table even though there was no LegacyContactInfo for those items. however, getOldLabels was not upgraded to use ContactInfo. so when getting old labels, no LegacyContactInfo was found to represent the origin of some crds values, and some of those items are removed from the crds table.

this occasionally causes double frees, segfaults, and panics due to bincode.free being called in `remove`. this is rooted in some longstanding issues with how crds value memory is managed.

the fix here is to use ContactInfo (in addition to LegacyContactInfo) during getOldLabels. this means we can find the actual wallclock for every node, which prevents the unnecessary trimming that triggers these errors. the errors are still theoretically possible when removing items, and the root cause still needs to be addressed