fix(lru): locks were not locking the state
adds two tests that would have failed, but now pass.

previously the deferred unlock would be executed at the end of the if block, which was actually before the code that needed to be locked. so the lock wasn't doing anything. this makes it so the lock is not unlocked until the end of the function, by putting the defer statement in the function scope instead of a nested block.