fix(gossip): panic: switch on corrupt value - SocketAddr from ContactInfo during buildPullRequests #96
This fixes the issue where gossip panics during buildPullRequests because it got an invalid ContactInfo with a SocketAddr containing an invalid tag. The cause is that the buf returned from getContactInfos is not filled contiguously. some buf indices are skipped because the item is too old, and it is excluded by an if statement.

The fix is to track the source and target indices separately. only push the target index if an item was actually inserted.