fix(gossip): panic: switch on corrupt value - SocketAddr from ContactInfo during buildPullRequests
fixes #96

This fixes the issue where gossip panics during buildPullRequests because it got an invalid ContactInfo with a SocketAddr containing an invalid tag.

The cause here is different from the main branch. The buf is filled, but it is returned with the wrong length. The length may be less than size if some items were filtered out.

The fix is to count how many items are actually included, and then return that as the len, instead of attempting to pre-calculate the len with insufficient information