fix(gossip): index out of bounds in filterBasedOnShredVersion
the index was incrementing +1 on each iteration even though the slice was shrinking on some iterations. so it would eventually get to an index that was no longer valid

also it was skipping some values because it would increment after swapRemove. the new item placed in that slot would not be checked.

this change increments the index only when an item is not removed. and it always gets the item based on the index, rather than always jumping to the next item