fix(remove_unused.py): was removing lines used in other imports
sometimes an import is only used as a shortcut for other imports. this addresses this to make sure lines are not used in other imports

this also addresses a related issue. if an import is *only* used in another import, and that other import is unused, then both should be removed. this accomplishes it by repeating the process until an iteration completes without removing any lines. i believe this is how brennan originally had the logic but it was lost when i rewrote it