refactor(collections,blockstore): SortedMap + SharedHashMapDB
Improve SortedMap flexibility so that it can be customized and has auto logic to work with slice keys. Also adjust the configuration approach so SortedSet can inherit the same flexibility without any boilerplate.

Use SortedSet as the storage backend for SharedHashMapDB in blockstore to support an iterator feature which is needed