refactor(blockstore): Support the Database abstraction without requiring it to be a middleman.
This means we can define the database directly as:
```zig
const BlockstoreDB = RocksDB;
```
instead of this:
```zig
const BlockstoreDB = Database(RocksDB);
```
... and critically, this change works to make this possible *without* losing most of the benefits of the Database abstraction.

The original approach of composition/delegation is still usable, but the first approach is now the one that is in use.

This change also moves some things around to improve clarity/consistency. Write batch naming is unified on `WriteBatch` as the struct and `writeBatch` as the method, and the struct is defined directly in the database to avoid too many layers of type aliases.

## Details

Previously, blockstore was set up to use the abstract Database struct as the database itself, which then is implemented by simply delegating out to the actual underlying implementation. This is how the trait/interface pattern is typically implemented in zig (see Allocator, Writer, Random, etc.). The advantage is that it guarantees that any code using the Database is generic and only depends on the abstraction, and it guarantees that any database implementation conforms to the same spec. This means that the code is more modular, and every component is very explicit about its minimal set of dependencies, which makes it easier to test and refactor the code. The disadvantage is that it introduces a layer of indirection which makes it harder to follow the concrete control flow of a program when viewing it holistically.

This change introduces a new approach for the trait/interface pattern. Instead of delegation through the interface struct, we have some comptime code that verifies that the implementation shares the same interface as the abstract struct. It does this off to the side, rather than by using Database as a middleman. Then, you have a new option for structs that depend on Database: use the concrete implementation directly in your code, while also having the assurance that it conforms to the abstraction. This makes it easier to follow the control flow of the program, with fewer searches and click-throughs to understand the implementation details.