fix(blockstore): merkle proof bug that was failing recovery test
there was a memory error here where the data behind pointers was being corrupted. it caused the recovery test to fail and then i wrote another merkle proof test (committed here) that reproduces it. i narrowed it down to be an issue where some data behind a pointer gets invalidated as soon as you call the gpa to allocate a totally separate and unrelated region of memory. this actually looks like a bug in the gpa so my fix was to stop using those pointers and to instead just deal with a fully owned slice. it doesn't directly address the memory corruption but it avoids it