fix(ledger): shred recovery (#241)
This PR includes a bug fix described in **Problem** as well as some related unit tests and refactoring changes that are described in **Other Changes**.

# Problem

On main, there is a bug in shred recovery where we see error.InvalidMerkleProof. It's thrown here:
https://github.com/Syndica/sig/blob/09bc1c9b5e27659dd545ebaef717f36886909aed/src/ledger/recovery.zig#L330-L331

This prevents most shreds from being recovered. Occasionally some shreds are recovered successfully, but usually we see this error preventing recovery.

## Background

Shred recovery uses erasure coding to recover shreds that were not received. For example, an erasure set may include 5 data shreds and 10 code shreds. The data shreds contain the information we care about. The code shreds only contain redundant data that allows us to recover the data shreds. From those 15 shreds, we could receive any 5 of them, and the reed solomon algorithm allows us to recover all 15 of the shreds.

Each erasure set also has a merkle tree calculated for it. Each shred contains its merkle proof.

During shred recovery, we calculate the merkle proof for every shred in the erasure set. For newly recovered shreds, the calculation is used to set the actual merkle proof in the shred, since that data is external to the erasure coding, and is not recovered by reed solomon.

We also calculate the merkle proof for every shred that we started with and already have the proof for. This is just a validation to ensure we are able to reproduce the same merkle proof that we already know is correct for that shred. This check ensures the integrity of the entire erasure set.

## Root cause

The error that we're seeing indicates that the wrong merkle proof was being calculated for existing shreds. The merkle proof was wrong because code shreds had the wrong indices being set. But the index field and the position in the common and code headers were being set incorrectly. Normally, each code shred should have a different index, representing its location in the erasure set and the slot. Instead, every code shred had an index that indicated it was the first shred in the erasure set.

I already added this logic to increment the index forward for each coding shred. As you can see it is assigned for this_common_header and this_coding_header. The bug is that those newly created headers with incremented indices are not actually being used.
https://github.com/Syndica/sig/blob/8637c2122d0242331c59de37c5102329bd97dc14/src/ledger/shredder.zig#L201-L209

## Fix

The fix is to use the headers with the incremented indices.
https://github.com/Syndica/sig/blob/09bc1c9b5e27659dd545ebaef717f36886909aed/src/ledger/recovery.zig#L279-L287

# Other changes

I made some other changes in the process of this investigation and fix.

## Shred recovery refactoring + tests
- move from shredder.zig into recovery.zig
- break apart massive `recover` function into several smaller functions
- unit test each function used for recovery. This also includes functions in shred.zig and reed_solomon.zig

## Rename "coding" shred names to "code"
Previously there was a mix of items named "code shred" and "coding shred". This change settles on "code shred" as the name to use everywhere. They are interchangeable so consistency is the priority here.

I also renamed variables and fields called "shred_variant" to "variant" when they are in the scope of a shred

## base58 refactoring
Hash, Pubkey, and Signature each need the same base58 logic for encoding statically sized arrays. I extracted out the common logic and put it into crypto/base58.zig

## eql function
I fleshed out the sig.utils.types.eql function to cover most of the cases I was looking for. This differs from std.meta.eql in that the std version means "are these actually the same exact item" and the sig version means "are these items equivalent". I believe the sig version is much more useful in most cases, whereas the std version is only useful in rare cases when you want to do things like comparing pointer equality. Specifically, the sig function:
https://github.com/Syndica/sig/blob/09bc1c9b5e27659dd545ebaef717f36886909aed/src/utils/types.zig#L280-L283