fix(geyser): GeyserWriter RecycleFBA pointers
previously the init function was broken because it made a copy of the allocator and contained a pointer to the original copy which was discarded from the stack

also, even after fixing that, when GeyserWriter would contain a mutable pointer to itself. the RecycleFBA should really be allocated on the stack. when the data is actually a part of the GeyserWriter struct, it means you have a struct that may be "const" but is still mutable because it contains a mutable pointer to itself. this violates aliasing assumptions and maybe it can trigger UB?