feat(core,io,ledger,bincode): test blockstore, serialize transactions, bincode config, PeekableReader (#251)
## ledger
The primary purpose of this PR is to add test coverage for the end-to-end blockstore behavior from inserting shreds to retrieving blocks. This is handled by a single test case that covers ShredInserter and BlockstoreReader.
https://github.com/Syndica/sig/blob/8c356f985aa6deece235618572fdd911d3fd3d9a/src/ledger/tests.zig#L48

In support of this, I have done some refactoring of the test helper code in the ledger package to make it less opinionated.

The code under test had many memory leaks that were revealed by this test. Otherwise there were no bugs detected, and getting blocks appears to works fine.

## bincode

Bincode already allows you to configure individual fields within a struct, and we make heavy use of that. Another feature was only partially supported by bincode. You can define within a struct how it should be configured (not its individual fields, but itself in entirety). In main this is only supported for the serialization of enums. I've extended this to be supported for serialization or deserialization of any types that support decls.

## core

To test the blockstore, I needed to load some entries and transactions that are generated by an analogous agave test. So I moved things around and added support for serialization.

I moved Entry out of ledger/reader.zig into core/entry.zig because it is a critical and widely used data structure in solana. Agave actually dedicates an entire crate to this data structure.

I also added tests to ensure Entry (and likewise all of its nested types) can be properly serialized and deserialized in a way that is consistent with agave. This includes VersionedTransaction, which is important to deserialize correctly if we want sig to process any transactions. Specifically I needed to implement a custom serializer for VersionedMessage that checks to see if the first bit is set to 1. If so, the remaining bits are a VersionedMessage according to the union's tag. If not, that bit and the remaining bits are recombined and deserialized as the legacy Message.

## io
To support the serialization of VersionedMessage, I added PeekableReader to utils/io.zig. Here's the general idea:

https://github.com/Syndica/sig/blob/8c356f985aa6deece235618572fdd911d3fd3d9a/src/utils/io.zig#L7-L17

This was useful for VersionedMessage because I wanted to peek the first byte but then potentially reuse it later for deserialization of other structs. I couldn't use a buffered reader for this because the buffered reader would pull arbitrary number of bytes out of the underlying reader, which means that we could potentially extract too many bytes out of the reader that was passed to this function. This would corrupt the calling scope and it would not be able to deserialize any more data.