fix(shred-collector): not processing shreds
there was a bool being used to decide whether to pass packets to the shred verifier, and it was being negated too many times. i fixed the logic.

i think the bug was introduced due to having the logic spread across too many scopes, which made it confusing. i tried to make it clearer by moving the logic into a narrower scope