fix(trace): require comptime message for log to avoid lifetime issues
this makes the code less flexible but the only alternative that is memory safe would be to dupe everything. dupe would be a waste since currently the strings are always comptime, so there's actually no need to dupe them. it's better to just require comptime for now and deal with allowing dynamic strings later if we need it.