fix(trace): incorrectly logging string fields (#308)
This fixes a logic bug that I introduced in the logger pr. The bug is that string fields are formatted incorrectly. []const u8 ends up being formatted as {any} instead of {s}.

This branch of the if-else will actually never execute because it has a redundant condition with the first branch. It's just not supposed to have that condition. I think I copy and pasted it by mistake.