refactor(ledger): rename "next_slots" to "child_slots" for consistency with "parent_slot", and clarity about the relationship
it's not just the "next" slots but the slots derived from this one, meaning they are specifically the child slots. there may be slots coming after this which are not derived from it, and they are excluded from this list. they are not children, even though they are technically "next". so child is a more accurate name