refactor(ledger): renames and docs for readability (#338)
* docs(ledger): transaction status field docs

* refactor(ledger): rename "parent offset" to "parent slot offset"

* refactor(ledger): rename shred field "signature" to "leader signature"

* refactor(ledger): rename CodeHeader's position to erasure_code_index

* docs(ledger): add doc comment for each column family in the schema

* refactor(ledger): rename "next_slots" to "child_slots" for consistency with "parent_slot", and clarity about the relationship

it's not just the "next" slots but the slots derived from this one, meaning they are specifically the child slots. there may be slots coming after this which are not derived from it, and they are excluded from this list. they are not children, even though they are technically "next". so child is a more accurate name

* refactor(ledger): rename BlockstoreWriter to LedgerResultWriter since it is responsible for storing the results of executing ledger items

* refactor(ledger): rename fec_set_index to erasure_set_index

abbreviations usually bad because they are too esoteric. there is already one word that has the same meaning. also that one word is used everywhere else, so it's more consistent to land on that single name.

* refactor(ledger): rename nextFecSetIndex to nextErasureSetIndex

* refactor(ledger): rename variable fec_set_size to erasure_set_size