fix(ledger): backward merkle root chaining bugs
It was ignoring result of checking chaining, and just always marked it as a duplicate (as if the check failed), and always returned true (as if the check succeeded). complete nonsense

It was comparing chained root to chained root, which is wrong. it should be chained to actual root

This change also refactors common logic into a single place.

Also this now returns an error if the merkle root is missing, because that means there is a severe bug in the validator and it needs to be shutdown and patched immediately.