feat(ledger): bench getting code shreds (#354)
* fix(ledger): backward merkle root chaining bugs

It was ignoring result of checking chaining, and just always marked it as a duplicate (as if the check failed), and always returned true (as if the check succeeded). complete nonsense

It was comparing chained root to chained root, which is wrong. it should be chained to actual root

This change also refactors common logic into a single place.

Also this now returns an error if the merkle root is missing, because that means there is a severe bug in the validator and it needs to be shutdown and patched immediately.

* Update src/ledger/shred_inserter/merkle_root_checks.zig

Co-authored-by: InKryption <59504965+InKryption@users.noreply.github.com>

* fix(ledger): zig fmt

* refactor(ledger): merkle root validator struct

the four methods all have basically the same dependencies so this bundles them together to clarify their purpose and simplify their expression

* refactor(ledger): use working stores to clarify and minimize dependencies, and to organize working state code

* refactor(ledger): in merkle root checks, rename early and late shreds to old and new shreds

* Added code shreds

* Add BlockstoreReader.getCodeShred

* Bench reader.getCodeShredsForSlot

---------

Co-authored-by: Drew Nutter <dnut@users.noreply.github.com>
Co-authored-by: InKryption <59504965+InKryption@users.noreply.github.com>