perf(shred-collector): cache verified merkle roots (#376)
This improves the performance of signature verification in the shred collector by a factor of up to 64.

The shred signature signs the merkle root. Since all the shreds from an erasure set share the same merkle root, you actually only need to verify the signature once for that merkle root, then every shred from the erasure set should be considered "verified" as long as you can reproduce the merkle root from its merkle proof.

Before this pr, the code was redundantly verifying the signature of the merkle root for every single shred. This leads to a performance bottleneck in the shred verifier because signature verification is computationally expensive.

This change removes the redundancy. It uses an lru cache to keep track of the last 1024 merkle roots to be verified.