perf(bechmarks): use direct print logger instead of channel logger (#425)
When I look at the flamegraphs from running benchmarks, the majority of the time is spent in the logger, busy spinning while waiting for log messages. This makes it harder to understand benchmark results.

I don't see a reason to use something as complex as a channel logger just to print information about the benchmarks. So I switched it to the logger that simply prints in the current thread.

Seeing these flamegraphs makes me concerned about the channel still being overly greedy of system resources. But I'm not trying to address this in the current pr. I have another PR which adds blocking recv methods to the channel to prevent busy spinning.