fix(ledger): improve database reliability with hashmap db (#400)
This switches the default database from rocksdb to the hashmap db.

To get the hashmap db working reliably over extended periods of time, it uses the disk backed allocator. Without this, the entire database would be stored in memory, eventually resulting in OOM.

To avoid creating absurd quantities of files that would cause errors due to too many open files, the disk allocator is only used occasionally to allocate 1 GB all at once. This 1 GB is consumed gradually over time. I implemented a new allocator called BatchAllocator that is responsible for executing this large batching. It acts as an intermediary between the disk allocator and the database. It provides an ordinary allocator interface which is used by the hashmap db.

To get parity with rocksdb in the benchmarks, I added reference counting in the database. When an entire shred needs to be read out of the hashmap db, previously it allocated 1 KB and copied the shred. Now it just copies the pointer out of the database, and it uses reference counting to ensure that the pointer lives sufficiently long.