fix(ledger): avoid rocksdb segfault (#438)
A segfault occurs in RocksDB when built in ReleaseSafe. This can be reproduced with a minimal example of code that only depends on the C library. It does not segfault when RocksDB is built using cmake. It does not segfault in any other zig optimize mode.

This PR is a quick fix to avoid the segfault by preventing RocksDB from being built by zig with ReleaseSafe. Whenever ReleaseSafe is specified as the build mode, RocksDB will be built with ReleaseFast. Otherwise it will use the same optimize mode as the build as a whole.