fix(ci): kcov dependencies on workflow runner (#467)
On the github runner, kcov is looking for libraries with version 38, but the system only has version 42. To get this working, I modified the workflow to hard link the files from 42 to 38. It's not an elegant fix, but currently all of our builds are failing, and this gets things working again. I created a ticket to look into this further: https://github.com/orgs/Syndica/projects/2/views/10?pane=issue&itemId=91539549